Create `Ledger::Mapper`
