+++ spec/models/ledger/item_spec.rb
@@ -310,71 +310,6 @@ RSpec.describe Ledger::Item, type: :model do
     end
   end
 
-  describe "#map!" do
-    let(:item) do
-      i = Ledger::Item.new(amount_cents: 1000, memo: "Test", date: Time.current)
-      i.save(validate: false)
-      i
-    end
-
-    it "returns nil when no event or card grant can be calculated" do
-      expect(item.map!).to be_nil
-      expect(item.primary_ledger).to be_nil
-    end
-
-    context "when an event can be calculated" do
-      it "creates a primary ledger and mapping for the event" do
-        event = create(:event)
-        allow(item).to receive(:calculate_card_grant).and_return(nil)
-        allow(item).to receive(:calculate_event).and_return(event)
-
-        item.map!
-        item.reload
-
-        expect(item.primary_ledger).to be_present
-        expect(item.primary_ledger.primary?).to be true
-        expect(item.primary_ledger.event).to eq(event)
-      end
-    end
-
-    context "when a card grant can be calculated" do
-      it "creates a primary ledger and mapping for the card grant" do
-        event = create(:event, :with_positive_balance)
-        card_grant = create(:card_grant, event:)
-        allow(item).to receive(:calculate_card_grant).and_return(card_grant)
-
-        item.map!
-        item.reload
-
-        expect(item.primary_ledger).to be_present
-        expect(item.primary_ledger.primary?).to be true
-        expect(item.primary_ledger.card_grant).to eq(card_grant)
-      end
-    end
-
-    it "reuses an existing ledger for the same event" do
-      event = create(:event)
-      existing_ledger = Ledger.create!(primary: true, event:)
-
-      allow(item).to receive(:calculate_card_grant).and_return(nil)
-      allow(item).to receive(:calculate_event).and_return(event)
-
-      item.map!
-      item.reload
-
-      expect(item.primary_ledger).to eq(existing_ledger)
-    end
-
-    it "is idempotent" do
-      event = create(:event)
-      allow(item).to receive(:calculate_card_grant).and_return(nil)
-      allow(item).to receive(:calculate_event).and_return(event)
-
-      item.map!
-      expect { item.map! }.not_to(change { Ledger::Mapping.count })
-    end
-  end
-
   describe "#calculate_amount_cents" do
     let(:item) do
       i = Ledger::Item.new(amount_cents: 0, memo: "Test", date: Time.current)
